Plug a GtkTreeModelFilter inside HildonTouchSelectorColumn
In order to use the live search widget in the touch selector, it is
necessary to separate the actual model with the data from the one in
the view, through a GtkTreeModelFilter. This adds a GtkTreeModelFilter
to HildonTouchSelectorColumn, and updates all the API to transparently
access the GtkTreeModelFilter iterators and paths whenever necessary.