Use GtkTreePaths instead of OssoABookContact pointers in the selection map
Make use of the uniqueness of the GtkTreePaths for each row in the
base model to replace the OssoABookContact pointers used in the row
selection map.

For this, add a new hash function to create the hash from the string
representation of a GtkTreePath. Also, add a key comparison function
using gtk_tree_path_compare().