Add live search capabilities to HildonTouchSelector
This adds a construct-only, :live-search boolean property to HildonTouchSelector.
When enabled, a HildonLiveSearch widget will be placed below the pannable area in
each column of the touch selector.

Also, the CAN_FOCUS flag for the treeviews in the touch selector will only be
unset for widgets without live search capabilities. CAN_FOCUS is necessary
to make sure that we get the input focus necessary for the live search
to work.