Fix HildonLiveSearch with multiple models in the view
Don't assume that the GtkTreeModel given to HildonLiveSearch
is the one used by the GtkTreeView. hildon-fm uses a
GtkTreeModelSort after the GtkTreeModelFilter, so it makes
the necessary conversions of iterators if the application has a
GtkTreeModelSort.

Don't assume that items in the view have a deepness of 0.
In selection_map_update_selection_from_map and
selection_map_update_map_from_selection, it
makes the conversions of GtkTreePath.

Fixes: NB#140251, Livesearch for File Manager

Signed-off-by: Claudio Saavedra <csaavedra@igalia.com>