Don't filter asynchronously when hildon_live_search_set_text() called
This is because calls to hildon_live_search_set_text() followed by a
retrieval of the selection will then break, and at least addressbook
assumes that this will happen synchronously.

Since we only want to give better responsiveness in the case when the
user is typing, no need to go fully asynchronously.