Fix hildon_pannable_area_add.
Calling gtk_container_add is not the correct way to chain
up to GtkBin's container_class->add. This just causes
hildon_pannable_area_add to be called again and we're in
an infinite loop. In-tree widgets have the lovely convenience
function _gtk_bin_add_child but we can't call that so we have
to jump through these hoops.